Added GUI immediately visible after program startup. Allows the user to login, and choose Lecture to join. Only a tutor can begin a lecture, if a user attempts to join a lecture without a tutor, he/she will be immediately disconnected.
However, the functions used for logging in/selecting lectures are dummy functions, which should be replaced with actual SOAP Web Operation calls.

Details about user and lecture are now in the UserBean and LectureBean classes. The LectureBean object is required for starting a channel.
Added a getLecture method to TeleChannel.